Remove leading part of reported file paths for MRI data Juelich
This removes the leading part of the file paths, leaving us with a
flat hierarchy of tarballs in the top level directory. This matches
what the data looks like in the Z03 folder in the FZJ ICF server.

The paths were initially mailed in as "Z03/P*.tar", and then got
changed to an artificial "sourcedata/DICOM/P*.tar", to mirror the
then-available paths for the Cologne dataset (sourcedata/NIfTI/*).

However, the Cologne dataset had since received an update (also mailed
in, cc76385) which changed the organisation there, doing away with the
"sourcedata/NIfTI" part. Moreover, the "sourcedata" designation,
although present in BIDS and suggested for DICOM files, is arbitrary
anyway.

We looked at this once again, and decided that we do not want to
introduce artificial levels, and the top-level "Z03" (or any other)
directory is redundant for the DICOM dataset ("the rule is: whatever
the native representation is, and nothing on top").